+++ packages/entities/entities-plugins/sandbox/pages/CustomPluginFormPage.vue
@@ -15,7 +15,7 @@
       <CustomPluginForm
         :config="kongManagerConfig"
         :plugin-name="pluginName"
-        :unsupported-types="['installed']"
+        :unsupported-types="['installed', 'streamed']"
       />
     </div>
   </div>