+++ packages/entities/entities-plugins/src/constants/index.ts
@@ -5,6 +5,7 @@ export const FEATURE_FLAGS = {
   KM_2306_CONDITION_FIELD_314: 'KM-2306-condition-field-314',
   KM_2446_DATAKIT_JWT_NODES: 'KM-2446-Datakit-JWT-nodes',
   KM_2503_CUSTOM_PLUGIN_FREEFORM: 'KM-2503-custom-plugin-freeform',
+  KM_2485_CLONED_PLUGINS: 'KM-2485-cloned-plugins',
 }
 
 export const TOASTER_PROVIDER = Symbol('TOASTER_PROVIDER')