+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.9.47](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@8.9.46...@kong-ui-public/dashboard-renderer@8.9.47) (2026-04-28)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
 ## [8.9.46](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@8.9.45...@kong-ui-public/dashboard-renderer@8.9.46) (2026-04-27)
 
 **Note:** Version bump only for package @kong-ui-public/dashboard-renderer