+++ packages/entities/entities-consumer-credentials/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.23](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@3.7.22...@kong-ui-public/entities-consumer-credentials@3.7.23) (2026-04-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
 ## [3.7.22](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@3.7.21...@kong-ui-public/entities-consumer-credentials@3.7.22) (2026-04-26)
 
 **Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials