+++ packages/entities/entities-consumers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.23](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.3.22...@kong-ui-public/entities-consumers@4.3.23) (2026-04-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumers
+
+
+
+
+
 ## [4.3.22](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.3.21...@kong-ui-public/entities-consumers@4.3.22) (2026-04-26)
 
 **Note:** Version bump only for package @kong-ui-public/entities-consumers